Refactor user detail page to show user info and addresses
Replaces the order detail view in userdetail/page.tsx with a user profile view, displaying user information and a list of shipping addresses. Updates the corresponding CSS for the new layout and styles. Also updates the users list page to link user IDs and view buttons to the user detail page.